requirements.txt
Flask==2.2.0
gunicorn==20.1.0
openpyxl==3.0.10
PyPDF2==2.9.0
python-dotenv==0.20.0
xlrd==2.0.1